Add workarounds to enable multi-processor support

+++ framework/aster-frame/src/cpu.rs
@@ -306,7 +306,7 @@ pub unsafe fn bsp_init() {
     }
     // Safety: FS register is not used for any other purpose.
     unsafe {
-        set_cpu_local_base_addr(__bsp_local_start as u64);
+        set_cpu_local_base_addr(__bsp_local_start as usize as u64);
     }
 }
